llmnova-hydra-Gamma-1-main/requirements.txt
fastapi
uvicorn
websockets
python-multipart
python-dotenv
openai
pydantic
rich
watchdog
requests
beautifulsoup4
playwright
tenacity
anthropic
python-pptx>=0.6.23
reportlab>=4.0.0
Pillow>=10.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
prometheus-client>=0.19.0
psutil>=5.9.0
apscheduler
redis==5.0.1
pytest-cov
fpdf2